feat: APIの構成を整理し、定数をconfig.jsに追加
