+++ deploykit/test/result.txt
@@ -1,0 +1,20 @@
+{"success":true,"reason":["ユーザーが既に存在します"]}
+{"success":false,"reason":["パラメータが不正です"]}
+{"success":false,"reason":["パラメータが不正です"]}
+{"success":false,"reason":["パラメータが不正です"]}
+{"success":false,"reason":["ユーザーIDが不正です"]}
+{"success":false,"reason":["パスワードが不正です"]}
+{"success":false,"reason":["ユーザーIDが不正です","パスワードが不正です"]}
+{"success":true,"reason":[]}
+{"success":false,"reason":["ユーザーが既に存在します"]}
+{"success":false,"reason":["パスワードが正しくありません"]}
+{"success":false,"reason":["パスワードが正しくありません"]}
+{"success":true,"reason":[]}
+{"success":false,"reason":["新しいパスワードが不正です"]}
+{"success":true,"reason":[]}
+{"success":false,"reason":["パスワードが正しくありません"]}
+{"success":true,"reason":[]}
+{"success":false,"reason":"新しいユーザーIDが不正です"}
+{"success":false,"reason":["パスワードが正しくありません"]}
+{"success":true,"reason":[]}
+{"success":true,"reason":[]}
