+++ litter-app/routes/api/change_id.js
@@ -1,20 +1,10 @@
-
-async function change_id(req) {// id変更
-    result = { result: { success: true, reason: [] }, status: SUCCESS };
-    try {
-        await pool.query("UPDATE litter.users SET user_id = ? WHERE user_id = ?", [req.new_id, req.id]);
-        result.result.success = true
-    } catch (error) {
-        result.result.success = false;
-        result.result.reason.push("データ更新に失敗しました");
-        result.status = INTERNAL_SERVER_ERROR;
-    }
-    return result;
-}
+const express = require('express');
+const router = express.Router();
+const common = require('../common');
+const config = require('../config.js');
 
 
-
-app.post('/change_id', async (req, res) => {// ユーザーID変更
+router.post('/', async (req, res) => {// ユーザーID変更
     /*
     idと新しいユーザーIDを受け取り、ユーザーIDを変更する。
     
@@ -27,37 +17,38 @@ app.post('/change_id', async (req, res) => {// ユーザーID変更
     */
     // パラメータのチェック
     allowedParams = ['id', 'password', 'new_id']
-    const paramCheckResult = check_parameters(req.body, allowedParams);
+    const paramCheckResult = common.check_parameters(req.body, allowedParams);
     if (!paramCheckResult.result.success) {
         res.status(paramCheckResult.status).json(paramCheckResult.result);
         return;
     }
     // バリデーション
-    const validationResult = validation(req.body);
+    const validationResult = common.validation(req.body);
     if (!(validationResult.result.success)) {
         res.status(validationResult.status).json(validationResult.result);
         return;
     }
     // 認証
-    const authResult = await is_correct(req.body); // パスワードが正しいかどうかを認証
+    const authResult = await common.is_correct(req.body); // パスワードが正しいかどうかを認証
     if (!authResult.result.success) {
         res.status(authResult.status).json(authResult.result);
         return;
     }
     // 既にいるかどうかのチェック
-    const existResult = await is_exist(req.body.new_id);
+    const existResult = await common.is_exist(req.body.new_id);
     if (existResult.result.success) {
-        res.status(BAD_REQUEST).json({ success: false, reason: "ユーザーが既に存在します" });
+        res.status(config.BAD_REQUEST).json({ success: false, reason: "ユーザーが既に存在します" });
         return;
     }
     // 新idのバリデーション
-    if (!idValidPattern.test(req.body.new_id)) {
-        res.status(BAD_REQUEST).json({ success: false, reason: "新しいユーザーIDが不正です" });
+    if (!config.idValidPattern.test(req.body.new_id)) {
+        res.status(config.BAD_REQUEST).json({ success: false, reason: "新しいユーザーIDが不正です" });
         return;
     }
     // ユーザーID変更
-    const result = await change_id(req.body); // ユーザーID変更
+    const result = await common.change_id(req.body); // ユーザーID変更
     res.status(result.status).json(result.result);
     return;
 })
 
+module.exports = router;
