+++ litter-app/routes/api/change_name.js
@@ -1,22 +1,10 @@
+const express = require('express');
+const router = express.Router();
+const common = require('../common');
+const config = require('../config.js');
 
 
-
-
-async function change_name(req) {// 名前変更
-    result = { result: { success: true, reason: [] }, status: SUCCESS };
-    try {
-        await pool.query("UPDATE litter.users SET name = ? WHERE user_id = ?", [req.new_name, req.id]);
-        result.result.success = true;
-    } catch (error) {
-        result.result.success = false;
-        result.result.reason.push("データ更新に失敗しました");
-        result.status = INTERNAL_SERVER_ERROR;
-    }
-    return result;
-}
-
-
-app.post('/change_name', async (req, res) => {// 名前変更
+router.post('/', async (req, res) => {// 名前変更
     /*
     idと新しい名前を受け取り、名前を変更する。
     入力:
@@ -28,26 +16,28 @@ app.post('/change_name', async (req, res) => {// 名前変更
     */
     // パラメータのチェック
     allowedParams = ['id', 'password', 'new_name']
-    const paramCheckResult = check_parameters(req.body, allowedParams);
+    const paramCheckResult = common.check_parameters(req.body, allowedParams);
     if (!paramCheckResult.result.success) {
         res.status(paramCheckResult.status).json(paramCheckResult.result);
         return;
     }
     // バリデーション
-    const validationResult = validation(req.body);
+    const validationResult = common.validation(req.body);
     if (!validationResult.result.success) {
         res.status(validationResult.status).json(validationResult.result);
         return;
     }
     // パスワードが正しいかどうかを確認
-    const authResult = await is_correct(req.body); // パスワードが正しいかどうかを確認
+    const authResult = await common.is_correct(req.body); // パスワードが正しいかどうかを確認
     if (!authResult.result.success) {
         res.status(authResult.status).json(authResult.result);
         return;
     }
     // 名前変更
-    const result = await change_name(req.body);
+    const result = await common.change_name(req.body);
     res.status(result.status).json(result.result);
     return;
 }
 )
+
+module.exports = router;
