+++ litter-app/routes/api/is_exist.js
@@ -1,23 +1,9 @@
-async function is_exist(value) {// ユーザーが存在するかどうかを確認
-    let result = { result: { success: true, reason: [] }, status: SUCCESS };
-    try {
-        const [rows] = await pool.query("SELECT id FROM litter.users WHERE user_id = ? and is_deleted = false", value);
-        if (rows.length > 0) {
-            result.result.success = true;
-        } else {
-            result.result.success = false;
-            result.result.reason.push("ユーザーが存在しません");
-        }
-    } catch (error) {
-        result.result.success = false;
-        result.result.reason.push("ユーザーが存在しません");
-        result.status = INTERNAL_SERVER_ERROR;
+const express = require('express');
+const router = express.Router();
+const common = require('../common');
+const config = require('../config.js');
 
-    }
-    return result;
-}
-
-app.post('/is_not_exist', async (req, res) => {
+router.post('/', async (req, res) => {
     /*
     idを受け取り、ユーザーが存在するかどうかを返す。
     入力:
@@ -27,14 +13,14 @@ app.post('/is_not_exist', async (req, res) => {
     */
     // パラメータのチェック
     const allowedParams = ['id']
-    const paramCheckResult = check_parameters(req.body, allowedParams);
+    const paramCheckResult = common.check_parameters(req.body, allowedParams);
     if (!paramCheckResult.result.success) {
         res.status(paramCheckResult.status).json(paramCheckResult.result);
         return;
     }
     // ユーザーが存在するかどうかを確認 
     // 本来ユーザが存在するか否かを返すべきだが、性質上｢いない｣場合に注目しているので逆転させている
-    const result = await is_exist(req.body.id);
+    const result = await common.is_exist(req.body.id);
     result.result.success = !result.result.success;
     result.result.reason = ["ユーザーが既に存在します"];
     res.status(result.status).json(result.result);
@@ -42,3 +28,4 @@ app.post('/is_not_exist', async (req, res) => {
 }
 )
 
+module.exports = router;
