+++ litter-app/routes/api/register.js
@@ -1,16 +1,9 @@
-async function register(req) {// ユーザー登録
-    result = { result: { success: true, reason: [] }, status: SUCCESS };
-    try {
-        await pool.query("INSERT INTO litter.users (user_id, name, password) VALUES (?, ?, ?)", [req.id, req.name, req.password]);
-        result.result.success = true;
-    } catch (error) {
-        result.result.success = false;
-        result.result.reason.push("データ挿入に失敗しました");
-        result.status = INTERNAL_SERVER_ERROR;
-    }
-    return result;
-}
-app.post('/register', async (req, res) => {// ユーザー登録
+const express = require('express');
+const router = express.Router();
+const common = require('../common');
+const config = require('../config.js');
+
+router.post('/', async (req, res) => {// ユーザー登録
     /*
     idとパスワードと名前を受け取り、ユーザーを登録する。
     既にユーザが存在している場合は失敗エラーを返す。
@@ -22,26 +15,27 @@ app.post('/register', async (req, res) => {// ユーザー登録
     }*/
     // パラメータのチェック
     allowedParams = ['id', 'password', 'name']
-    const paramCheckResult = check_parameters(req.body, allowedParams);
+    const paramCheckResult = common.check_parameters(req.body, allowedParams);
     if (!paramCheckResult.result.success) {
         res.status(paramCheckResult.status).json(paramCheckResult.result);
         return;
     }
-    // バリデーション
-    const validationResult = validation(req.body);
+    // 入力規則に合っているかチェック
+    const validationResult = common.validation(req.body);
     if (!validationResult.result.success) {
         res.status(validationResult.status).json(validationResult.result);
         return;
     }
     // 既にいるかどうかのチェック
-    const existResult = await is_exist(req.body.id);
+    const existResult = await common.is_exist(req.body.id);
     if (existResult.result.success) { // 既にいる場合
-        res.status(BAD_REQUEST).json({ success: false, reason: ["ユーザーが既に存在します"] });
+        res.status(config.BAD_REQUEST).json({ success: false, reason: ["ユーザーが既に存在します"] });
         return;
     }
     // ユーザー登録
-    const result = await register(req.body);
+    const result = await common.register(req.body);
     res.status(result.status).json(result.result);
     return;
 });
 
+module.exports = router;
