+++ litter-app/routes/api/remove.js
@@ -1,18 +1,10 @@
-async function remove(req) {// ユーザー削除
-    result = { result: { success: true, reason: [] }, status: SUCCESS };
-    try {
-        await pool.query("UPDATE litter.users SET is_deleted = true WHERE user_id = ?", [req.id]);
-        result.result.success = true;
-    } catch (error) {
-        result.result.success = false;
-        result.result.reason.push("データ削除に失敗しました");
-        result.status = INTERNAL_SERVER_ERROR;
-    }
-    return result;
-}
+const express = require('express');
+const router = express.Router();
+const common = require('../common');
+const config = require('../config.js');
 
 
-app.post('/remove', async (req, res) => {// ユーザー削除
+router.post('/', async (req, res) => {// ユーザー削除
     /*
     idとパスワードを受け取り、ユーザーを削除する。
     入力:
@@ -23,25 +15,27 @@ app.post('/remove', async (req, res) => {// ユーザー削除
     */
     // パラメータのチェック
     allowedParams = ['id', 'password']
-    const paramCheckResult = check_parameters(req.body, allowedParams);
+    const paramCheckResult = common.check_parameters(req.body, allowedParams);
     if (!paramCheckResult.result.success) {
         res.status(paramCheckResult.status).json(paramCheckResult.result);
         return;
     }
-    // バリデーション
-    const validationResult = validation(req.body);
+    // 入力規則に合っているかチェック
+    const validationResult = common.validation(req.body);
     if (!(validationResult.result.success)) {
         res.status(validationResult.status).json(validationResult.result);
         return;
     }
     // 認証
-    const authResult = await is_correct(req.body); // パスワードが正しいかどうかを確認
+    const authResult = await common.is_correct(req.body); // パスワードが正しいかどうかを確認
     if (!authResult.result.success) {
         res.status(authResult.status).json(authResult.result);
         return;
     }
     // ユーザー削除
-    const result = await remove(req.body);// ユーザー削除
+    const result = await common.remove(req.body);// ユーザー削除
     res.status(result.status).json(result.result);
     return;
 })
+
+module.exports = router;
