+++ note.md
@@ -152,3 +152,32 @@ VSCode を再起動
 -   IDの大文字小文字をDBに記録していない
     -   →Twiterは表示では区別しているものの、大小文字の違いはダブって登録することができない
 -   どう取り扱うか要件等
+
+2025-03-31 13:30:31
+-   apiファイル分割した！わかりにくくなった節もあるのでちょっと図解
+    ```
+    litter-app/routes/ 内
+
+      ┌──────────────┐  ┌──────────────┐  ┌──────────────┐
+      │  common.js   │  │  config.js   │  │    db.js     │
+      └───────┬──────┘  └───────┬──────┘  └───────┬──────┘
+              ▼                 ▼                 ▼
+              └─────────────────┴─────────────────┘
+                │                       │
+                ▼                       ▼              
+          ┌─────────────┐       ┌──────────────────┐
+          │   api.js    │ ◀───  │    ./api/*.js    │
+          └──────┬──────┘       │                  │
+                 ▼              └──────────────────┘
+          ┌──────────────┐
+          │    ユーザー   │
+          └──────────────┘
+
+    ```
+    -   ``common.js``は各apiが共通で使う機能をとりまとめたところ
+    -   ``config.js``はサーバー管理者用設定ファイル
+    -   `db.js`はdb設定ファイル
+    -   ``api.js``は各apiの振り分けを行う部分api.jsにはじめにやってきて、各apiごとに適切な関数を``./api/*.js``から呼び出す。
+    -   今後apiが増えたら``./api/*.js``を増やして対応する。
+
+
